Create README.md
Hi welcome to my personal website. Let me introduced myself i am have my personal since high school called Kenneth, a third year Business Administration Major in University of Merdeka Malang. I think of myself as a friendly and comminicative person, and also talk about business and marketing! I am passionate about utilizing the power of marketing and data to solve real world problems that can lead to the better life of many people in the world. With a strong foindation in data analyst and also in marketing learning, i am equipped with the skills and expertise to tackle a wide range of challenges. On this website, you Will be able to know more about me and my work, as well as my experiences in the marketing. I am inviting you to explore more about myself and my journey as a marketing enthusiast. Hit me up on WhatsApp so that can we areange some lovely talks. Thankyou for visiting